fix (flutter http caching page update):
 - update the http caching file
