+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: flutter_http_caching
-description: A new Flutter package project.
+description: Best way for caching http request
 version: 0.0.1
-homepage:
+homepage: https://github.com/varunsat/flutter_http_caching
 
 environment:
   sdk: '>=2.18.1 <3.0.0'
@@ -10,16 +10,18 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^2.0.0
   path_provider: ^2.0.11
   sqflite: ^2.1.0
   http_interceptor: ^1.0.2
   http: ^0.13.5
   internet_connection_checker: ^0.0.1+4
+  path: ^1.8.2
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^2.0.0
+
 
 
 # For information on the generic Dart part of this file, see the
